[refac] Refactor package structure for OpenAI integration
